Reorganize components structure
