+++ src/components/filters/ExtendedDateFilter/DynamicSelect/DynamicSelect.scss
